// todo_frontEnd/todo_frontend_tsc/src/typescript_client_latest/api.ts
const BASE_PATH = "https://f1lbive9qi.execute-api.us-east-1.amazonaws.com/dev".replace(/\/+$/, "");

/**
 *
 * @export
 */
export const COLLECTION_FORMATS = {
    csv: ",",
    ssv: " ",
    tsv: "\t",
    pipes: "|",
};

/**
 *
 * @export
 * @interface FetchAPI
 */
export interface FetchAPI {
    (url: string, init?: any): Promise<Response>;
}

/**
 *
 * @export
 * @interface FetchArgs
 */
export interface FetchArgs {
    url: string;
    options: any;
}

/**
 *
 * @export
 * @class BaseAPI
 */
export class BaseAPI {
    protected configuration?: Configuration;

    constructor(configuration?: Configuration, protected basePath: string = BASE_PATH, protected fetch: FetchAPI = isomorphicFetch) {
        if (configuration) {
            this.configuration = configuration;
            this.basePath = configuration.basePath || this.basePath;
        }
    }
};
// ...
export interface PostTodo {
    /**
     * 
     * @type {string}
     * @memberof PostTodo
     */
    name: string;
}
// ...
export interface UpdateTodo {
    /**
     * 
     * @type {string}
     * @memberof UpdateTodo
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateTodo
     */
    status?: string;
}
// ...
export class TodoApi extends BaseAPI {
    /**
     * Add a new todo of an org
     * @summary Add a new todo
     * @param {PostTodo} body Create a new todo
     * @param {string} authToken login-token to authorize user
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof TodoApi
     */
    public addTodo(body: PostTodo, authToken: string, options?: any) {
        return TodoApiFp(this.configuration).addTodo(body, authToken, options)(this.fetch, this.basePath);
    }

    /**
     * delete a todo
     * @summary Deletes a todo
     * @param {string} authToken login-token to authorize user
     * @param {string} todoId todo id to delete
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof TodoApi
     */
    public deleteTodo(authToken: string, todoId: string, options?: any) {
        return TodoApiFp(this.configuration).deleteTodo(authToken, todoId, options)(this.fetch, this.basePath);
    }

    /**
     * delete a todo
     * @summary get a todo's history
     * @param {string} authToken login-token to authorize user
     * @param {string} todoId todo id for its history
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof TodoApi
     */
    public getTodoHistory(authToken: string, todoId: string, options?: any) {
        return TodoApiFp(this.configuration).getTodoHistory(authToken, todoId, options)(this.fetch, this.basePath);
    }

    /**
     * fetch todos
     * @summary fetch all todos
     * @param {string} authToken login-token to authorize user
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof TodoApi
     */
    public getTodos(authToken: string, options?: any) {
        return TodoApiFp(this.configuration).getTodos(authToken, options)(this.fetch, this.basePath);
    }

    /**
     * Update todo by Id
     * @summary Update todo by ID
     * @param {string} authToken login-token to authorize user
     * @param {string} todoId ID of todo to update
     * @param {UpdateTodo} [body] update a created todo
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof TodoApi
     */
    public updateTodo(authToken: string, todoId: string, body?: UpdateTodo, options?: any) {
        return TodoApiFp(this.configuration).updateTodo(authToken, todoId, body, options)(this.fetch, this.basePath);
    }

}